requirements.txt
streamlit==1.26.0
snowflake-snowpark-python==1.22.1
snowflake.core==1.0.0
pandas==2.0.3
